Add libpmemkv_json helper library
Fixes: pmem/pmemkv#409
