+++ INSTALLING.md
@@ -27,10 +27,10 @@ Building from Sources
 
 * 64-bit Linux (OSX and Windows are not yet supported)
 * [PMDK](https://github.com/pmem/pmdk) - Persistent Memory Development Kit
-* [RapidJSON](https://github.com/tencent/rapidjson) - JSON parser
 * [libpmemobj-cpp](https://github.com/pmem/libpmemobj-cpp) - C++ bindings for PMDK (required by all engines except blackhole and caching)
 * [TBB](https://github.com/01org/tbb) - Thread Building Blocks (requiered by cmap & vcmap engines)
 * [memkind](https://github.com/memkind/memkind) - Volatile memory manager (required by vsmap & vcmap engines)
+* (optional) [RapidJSON](https://github.com/tencent/rapidjson) - JSON parser (required by `libpmemkv_json_config` helper library)
 * Used only for development & testing:
 	* [pandoc](https://pandoc.org/) - markup converter to generate manpages
 	* [doxygen](http://www.doxygen.nl/) - tool for generating documentation from annotated C++ sources
@@ -57,6 +57,20 @@ ctest --output-on-failure
 ```
 
 to see the output of failed tests.
+
+Building of the `libpmemkv_json_config` helper library is enabled by default.
+If you want to disable it (for example to get rid of the RapidJSON dependency)
+run:
+
+```sh
+cmake .. -DBUILD_JSON_CONFIG=OFF
+```
+
+instead of:
+
+```sh
+cmake ..
+```
 
 **Managing shared library**
